🚚📚 Move: all dynamic script files to `$lib/scripts` directory
- 🚚 Move: `$lib/*.ts` to `$lib/scripts/` directory
- 🚚 Move: `$lib/i18n` directory to `$lib/scripts/` directory
- 🚚 Move: `$lib/data` directory to `$lib/scripts/` directory
- 📚 Docs: update docs
