+++ src/lib/components/RevealFadeIn.svelte
@@ -2,7 +2,7 @@
 <script lang="ts">
 	import { onMount } from 'svelte';
 	import { browser } from '$app/environment';
-	import { addClassOnVisible } from '$lib/util';
+	import { addClassOnVisible } from '$lib/scripts/util';
 
 	let effect: HTMLElement;
