+++ src/lib/components/news/NewsList.svelte
@@ -3,7 +3,7 @@
 	import ChevronArrow from './util/ChevronArrow.svelte';
 	import ArticleCard from './ArticleCard.svelte';
 
-	import type { ArticleMetadata, ArticleThumbnailImgFmts } from '$lib/types';
+	import type { ArticleMetadata, ArticleThumbnailImgFmts } from '$lib/scripts/types';
 	import { fly } from 'svelte/transition';
 
 	export let articles: ArticleMetadata[];
