+++ src/lib/scripts/data/SPONSORS.ts
