+++ src/lib/scripts/i18n/index.ts
