+++ src/lib/scripts/i18n/locales/ko.json
