+++ src/routes/+page.svelte
@@ -9,7 +9,7 @@
 	import Sponsors from './Sponsors.svelte';
 
 	import type { PageData } from './$types';
-	import { SITE_URL } from '$lib/variables';
+	import { SITE_URL } from '$lib/scripts/variables';
 	import { _ } from 'svelte-i18n';
 
 	export let data: PageData;
