+++ src/routes/+page.ts
@@ -1,7 +1,7 @@
 // © 2022 REVATI
 
 import type { PageLoad } from './$types';
-import type { ArticleMetadata, ArticleThumbnailImgFmts } from '$lib/types';
+import type { ArticleMetadata, ArticleThumbnailImgFmts } from '$lib/scripts/types';
 
 export const load: PageLoad = async ({
 	fetch,
