+++ src/routes/news/articles/[slug]/+page.ts
@@ -2,7 +2,7 @@
 
 import type { PageLoad } from './$types';
 import type { SvelteComponent } from 'svelte';
-import type { ArticleMetadata } from '$lib/types';
+import type { ArticleMetadata } from '$lib/scripts/types';
 import { error } from '@sveltejs/kit';
 
 export const load: PageLoad = async ({
